refactor contributions tables
